added conio for linux, added exception classes, added keyboard interrupts, added kernelmode checks, added eret instruction

+++ MIPS-VM/helper.h
@@ -20,7 +20,7 @@ static To& bit_cast(const From& in) {
 }
 
 static bool string_terminates(const char* str, size_t buf_size) {
-    for (int i = 0; i < buf_size; i++) {
+    for (size_t i = 0; i < buf_size; i++) {
         if (str[i] == 0) {
             return true;
         }
